# api/openapi.yaml
openapi: 3.1.0
info:
  title: go-service-main API
  version: 0.4.0
  description: Single-tenant administration API compatible with vue3-element-admin.
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '400': { $ref: '#/components/responses/AdminAPIError' }
        '401': { $ref: '#/components/responses/AdminAPIError' }
  /api/v1/auth/refresh-token:
    post:
      operationId: refreshToken
      security: []
      parameters:
        - in: query
          name: refreshToken
          required: true
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '401': { $ref: '#/components/responses/AdminAPIError' }
...
  /api/v1/users:
    get:
      operationId: listUsers
      parameters:
        - { in: query, name: pageNum, schema: { type: integer, minimum: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { in: query, name: keywords, schema: { type: string } }
        - { in: query, name: status, schema: { type: integer, enum: [0, 1] } }
        - { in: query, name: deptId, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    post:
      operationId: createUser
      requestBody: { $ref: '#/components/requestBodies/UserForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '400': { $ref: '#/components/responses/AdminAPIError' }
        '403': { $ref: '#/components/responses/AdminAPIError' }
        '409': { $ref: '#/components/responses/AdminAPIError' }
...
  /api/v1/users/profile:
    get:
      operationId: getUserProfile
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    put:
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/users/password:
    put:
      operationId: changeCurrentUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '400': { $ref: '#/components/responses/AdminAPIError' }
  /api/v1/users/{userId}/form:
    parameters:
      - $ref: '#/components/parameters/UserID'
    get:
      operationId: getUserForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '404': { $ref: '#/components/responses/AdminAPIError' }
  /api/v1/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserID'
    put:
      operationId: updateUser
      requestBody: { $ref: '#/components/requestBodies/UserForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '404': { $ref: '#/components/responses/AdminAPIError' }
    delete:
      operationId: deleteUsers
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '403': { $ref: '#/components/responses/AdminAPIError' }
  /api/v1/users/{userId}/status:
    parameters:
      - $ref: '#/components/parameters/UserID'
      - { in: query, name: status, required: true, schema: { type: integer, enum: [0, 1] } }
    patch:
      operationId: updateUserStatus
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/users/{userId}/password/reset:
    parameters:
      - $ref: '#/components/parameters/UserID'
      - { in: query, name: password, required: true, schema: { type: string, minLength: 8 } }
    put:
      operationId: resetUserPassword
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/users/import:
    post:
      operationId: importUsers
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
        '400': { $ref: '#/components/responses/AdminAPIError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    UserID:
      in: path
      name: userId
      required: true
      schema: { type: string }
...
  requestBodies:
    UserForm:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UserForm' }
...
  responses:
    AdminAPISuccess:
      description: Successful administration API response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAPIResult' }
    AdminAPIError:
      description: Administration API error with a stable code.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAPIResult' }
  schemas:
    LoginRequest:
      type: object
      required: [username, password, captchaId, captchaCode]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
        captchaId: { type: string, minLength: 1 }
        captchaCode: { type: string, minLength: 1 }
    UserForm:
      type: object
      required: [username, nickname, status, roleIds]
      properties:
        id: { type: string }
        username: { type: string }
        nickname: { type: string }
        mobile: { type: string }
        gender: { type: integer, minimum: 0, maximum: 2 }
        avatar: { type: string }
        email: { type: string }
        status: { type: integer, enum: [0, 1] }
        deptId: { type: string }
        roleIds:
          type: array
          items: { type: integer, format: int64 }
    ProfileForm:
      type: object
      required: [nickname, gender]
      properties:
        nickname: { type: string }
        avatar: { type: string }
        gender: { type: integer, minimum: 0, maximum: 2 }
    PasswordChangeForm:
      type: object
      required: [oldPassword, newPassword, confirmPassword]
      properties:
        oldPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
        confirmPassword: { type: string, minLength: 8 }
...
    AdminAPIResult:
      type: object
      required: [code, msg, data]
      properties:
        code: { type: string }
        msg: { type: string }
        data: {}
        requestId: { type: string }